backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
redis==5.0.1
celery==5.3.4
django-storages==1.14.2
boto3==1.34.14
django-filter==23.5
drf-yasg==1.21.7
python-decouple==3.8
gunicorn==21.2.0
django-ffmpeg==1.0.5
djangorestframework-gis==1.0.0
django-ckeditor==6.7.0
django-import-export==3.3.6
django-debug-toolbar==4.2.0
sentry-sdk==1.39.1
whitenoise==6.6.0
django-allauth==0.57.0
phonenumbers==8.13.29
django-phonenumber-field==7.2.0
requests==2.31.0
stripe==7.7.0
flutterwave==1.2.0
